requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
emoji==1.2.0
flake8==3.9.1
idna==2.10
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytaf==1.2.1
python-dotenv==0.17.0
python-telegram-bot==13.4.1
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
